# engine-skeleton/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FinOps Watch \u2014 AI Engine",
    "description": "Production AI Engine for TF2 FinOps Watch. Detects AWS cost anomalies via per-mechanism statistical detectors, plans containment actions, and verifies remediation outcomes. Contract: ai-api-contract.md v1.5.0.",
    "contact": {
      "name": "TF2 FinOps Watch \u2014 AI Team",
      "email": "trankanjin803@gmail.com"
    },
    "license": {
      "name": "Internal \u2014 TF2 Capstone Phase 2"
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/v1/status/{anomaly_id}": {
      "get": {
        "tags": [
          "Status"
        ],
        "summary": "Get Remediation Status",
        "description": "Poll the current remediation status for a specific anomaly_id. If the tenant is in LOCKED_MODE, response headers will include X-Containment-Status: LOCKED.",
        "operationId": "get_status_v1_status__anomaly_id__get",
        "parameters": [
          {
            "name": "anomaly_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Anomaly ID in format ANM-YYYY-MMDD{A-Z}",
              "title": "Anomaly Id"
            },
            "description": "Anomaly ID in format ANM-YYYY-MMDD{A-Z}"
          },
          {
            "name": "X-Tenant-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Tenant-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RemediationStatusResponse"
                }
              }
            }
          },
          "403": {
            "description": "Anomaly belongs to a different tenant",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Anomaly ID not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded (100 req/min/tenant)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/audit/{audit_id}/rollback": {
      "post": {
        "tags": [
          "Status"
        ],
        "summary": "Record Manual Rollback (False Positive)",
        "description": "CDO notifies the AI Engine that a remediation was manually rolled back (e.g. false positive). Updates the false-positive counter for model feedback and recalculates the error budget.",
        "operationId": "record_rollback_v1_audit__audit_id__rollback_post",
        "parameters": [
          {
            "name": "audit_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Anomaly/Audit ID in format ANM-YYYY-MMDD{A-Z}",
              "title": "Audit Id"
            },
            "description": "Anomaly/Audit ID in format ANM-YYYY-MMDD{A-Z}"
          },
          {
            "name": "X-Tenant-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Tenant-Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RollbackRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RollbackResponse"
                }
              }
            }
          },
          "404": {
            "description": "Audit ID not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "ActionLogEntry": {
        "properties": {
          "timestamp": {
            "type": "string",
            "title": "Timestamp",
            "description": "RFC3339 UTC timestamp of this action"
          },
          "action": {
            "$ref": "#/components/schemas/ContainmentAction"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "description": "e.g. COMPLETED, DRY_RUN_COMPLETED, FAILED"
          },
          "actor": {
            "type": "string",
            "title": "Actor",
            "description": "Actor that executed the step"
          }
        },
        "type": "object",
        "required": [
          "timestamp",
          "action",
          "status",
          "actor"
        ],
        "title": "ActionLogEntry"
      },
...
      "ContainmentAction": {
        "type": "string",
        "enum": [
          "tag-for-review",
          "time-gated-countdown",
          "auto-shutdown",
          "quota-cap"
        ],
        "title": "ContainmentAction"
      },
...
      "ErrorResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "description": "Always false for error responses",
            "default": false
          },
          "error_code": {
            "type": "string",
            "title": "Error Code",
            "description": "Internal error code (e.g. ERR_INVALID_SCHEMA)"
          },
          "error_message": {
            "type": "string",
            "title": "Error Message",
            "description": "Human-readable error detail"
          },
          "correlation_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Correlation Id",
            "description": "Trace ID if available"
          }
        },
        "type": "object",
        "required": [
          "error_code",
          "error_message"
        ],
        "title": "ErrorResponse",
        "description": "Standard error envelope returned on 4xx/5xx responses.",
        "example": {
          "correlation_id": "9b1deb4d-3b7d-4bad-9bdd-2b0d7b3dcb6d",
          "error_code": "ERR_INVALID_SCHEMA",
          "error_message": "Field 'data_source_type' is required.",
          "success": false
        }
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RemediationStatus": {
        "type": "string",
        "enum": [
          "PENDING_APPROVAL",
          "IN_PROGRESS",
          "SUCCESS",
          "ROLLED_BACK",
          "ESCALATED"
        ],
        "title": "RemediationStatus"
      },
      "RemediationStatusResponse": {
        "properties": {
          "audit_id": {
            "type": "string",
            "title": "Audit Id",
            "description": "Audit-record identifier (UUID v4) \u2014 audit chain, telemetry \u00a716"
          },
          "anomaly_id": {
            "type": "string",
            "pattern": "^ANM-\\d{4}-\\d{4}[A-Z]$",
            "title": "Anomaly Id",
            "description": "Anomaly ID (ANM-YYYY-MMDD{A-Z}) \u2014 matches the path param"
          },
          "status": {
            "$ref": "#/components/schemas/RemediationStatus"
          },
          "containment_locked": {
            "type": "boolean",
            "title": "Containment Locked",
            "description": "True if tenant is in LOCKED_MODE (dry-run only)"
          },
          "error_budget_remaining_pct": {
            "type": "number",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Error Budget Remaining Pct",
            "description": "Error budget remaining (0\u2013100 %)"
          },
          "actions_log": {
            "items": {
              "$ref": "#/components/schemas/ActionLogEntry"
            },
            "type": "array",
            "title": "Actions Log",
            "description": "Ordered history of steps taken"
          }
        },
        "type": "object",
        "required": [
          "audit_id",
          "anomaly_id",
          "status",
          "containment_locked",
          "error_budget_remaining_pct",
          "actions_log"
        ],
        "title": "RemediationStatusResponse",
        "description": "Response body for GET /v1/status/{id}.",
        "example": {
          "actions_log": [
            {
              "action": "tag-for-review",
              "actor": "finops-ai-engine-role",
              "status": "DRY_RUN_COMPLETED",
              "timestamp": "2026-06-26T10:05:46Z"
            }
          ],
          "anomaly_id": "ANM-2026-0626A",
          "audit_id": "3fa85f64-5717-4562-b3fc-2c963f66afa6",
          "containment_locked": false,
          "error_budget_remaining_pct": 97.3,
          "status": "PENDING_APPROVAL"
        }
      },
...
      "RollbackRequest": {
        "properties": {
          "reason": {
            "type": "string",
            "title": "Reason",
            "description": "Why the rollback was triggered (e.g. False Positive)"
          },
          "rolled_back_by": {
            "type": "string",
            "title": "Rolled Back By",
            "description": "Email of engineer who initiated rollback"
          }
        },
        "type": "object",
        "required": [
          "reason",
          "rolled_back_by"
        ],
        "title": "RollbackRequest",
        "description": "Request body for POST /v1/audit/{audit_id}/rollback.",
        "example": {
          "reason": "False positive \u2014 instance is used for approved experiment",
          "rolled_back_by": "engineer@company.com"
        }
      },
      "RollbackResponse": {
        "properties": {
          "rollback_recorded": {
            "type": "boolean",
            "title": "Rollback Recorded",
            "description": "True if audit log was updated"
          },
          "false_positive_count_updated": {
            "type": "boolean",
            "title": "False Positive Count Updated",
            "description": "True if FP counter was incremented for model feedback"
          },
          "new_error_budget_burned_pct": {
            "type": "number",
            "minimum": 0.0,
            "title": "New Error Budget Burned Pct",
            "description": "Updated error budget burn percentage after this event"
          },
          "containment_locked": {
            "type": "boolean",
            "title": "Containment Locked",
            "description": "True if LOCKED_MODE was triggered by this rollback"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Human-readable system message"
          }
        },
        "type": "object",
        "required": [
          "rollback_recorded",
          "false_positive_count_updated",
          "new_error_budget_burned_pct",
          "containment_locked",
          "message"
        ],
        "title": "RollbackResponse",
        "description": "Response body for POST /v1/audit/{audit_id}/rollback.",
        "example": {
          "containment_locked": true,
          "false_positive_count_updated": true,
          "message": "Error budget exceeded 1%. Tenant switched to LOCKED_MODE \u2014 all containment actions will be dry-run only.",
          "new_error_budget_burned_pct": 1.5,
          "rollback_recorded": true
        }
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}